Correccion de errores modulos: Escuelas y Ciclos
Se agrega SPATIE para la creacion de roles, permisos para los usuarios. Los modulos de escuelas y ciclos ya manejan los permisos de usuarios

+++ resources/views/front-end/ciclos/index.blade.php
@@ -3,12 +3,14 @@
 @section('content')
 <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-2 pb-1 mb-1 border-bottom">
     <h1 class="h2" style="color: #607D8B">Ciclos Escolares</h1>
+    @can('ciclos.create')
     <div class="btn-toolbar mb-1 mb-md-0">
         <a href="{{ route('ciclos.create') }}" class="btn text-white" role="button" aria-pressed="true" style="background-color: #4CAF50">
             <i class="fas fa-plus"></i>
             Nuevo Ciclo
         </a>
     </div>
+    @endcan
 </div>
 
 <div class="row justify-content-center">
@@ -40,15 +42,21 @@
                             @endif
                         </td>
                         <td class="text-center">
-                            <a href="{{ route('ciclos.show', ['id' => $ciclo->id]) }}" class="btn bg-transparent btn-sm" role="button" title="Ver" aria-pressed="true">
-                                <i class="far fa-eye text-secondary"></i>
-                            </a>
-                            <a href="{{ route('ciclos.edit', ['id' => $ciclo->id]) }}" class="btn bg-transparent btn-sm" role="button" title="Editar" aria-pressed="true">
-                                <i class="fas fa-pencil-alt text-primary"></i>
-                            </a>
-                            <button type="button" class="btn bg-transparent btn-sm" onclick="deleteItem('{{ route('ciclos.destroy', $ciclo->id) }}')">
-                                <i class="far fa-trash-alt text-danger"></i>
-                            </button>
+                            @can('ciclos.show')
+                                <a href="{{ route('ciclos.show', ['id' => $ciclo->id]) }}" class="btn bg-transparent btn-sm" role="button" title="Ver" aria-pressed="true">
+                                    <i class="far fa-eye text-secondary"></i>
+                                </a>
+                            @endcan
+                            @can('ciclos.edit')
+                                <a href="{{ route('ciclos.edit', ['id' => $ciclo->id]) }}" class="btn bg-transparent btn-sm" role="button" title="Editar" aria-pressed="true">
+                                    <i class="fas fa-pencil-alt text-primary"></i>
+                                </a>
+                            @endcan
+                            @can('ciclos.destroy')
+                                <button type="button" class="btn bg-transparent btn-sm" onclick="deleteItem('{{ route('ciclos.destroy', $ciclo->id) }}')">
+                                    <i class="far fa-trash-alt text-danger"></i>
+                                </button>
+                            @endcan
                         </td>
                     </tr>
                 @endforeach
@@ -62,7 +70,6 @@
 @section('module_javascript')
 <!-- Archivo(s) javascript del modulo -->
 <script src="{{ asset('js/axios.js') }}" ></script>
-<script src="{{ asset('js/modules/ciclo.js') }}"></script>
 <script>
     $(document).ready(function() {
         /*
@@ -91,5 +98,47 @@
             } );
         } ).draw();
     } );
+
+    //Inicia funcion para eliminar el ciclo escolar seleccionado
+    function deleteItem(urlDelete) {
+
+        Swal.fire({
+            title: '¿Realmente desea eliminar este ciclo?',
+            text: "Si hizo clic por equivocación, presione Cancelar",
+            type: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            cancelButtonText: 'Cancelar',
+            confirmButtonText: 'Si, deseo eliminarlo'
+        }).then((result) => {
+
+            if (result.value) {
+
+                axios.delete(urlDelete)
+                    .then(function (response) {
+                        //La eliminación se realizo correctamente
+                        Swal.fire({
+                            title: 'Ciclo escolar eliminado',
+                            text: 'El ciclo se elimino correctamente. Presione Continuar',
+                            type: 'success',
+                            allowOutsideClick: false,
+                            showConfirmButton: true,
+                            confirmButtonColor: '#3085d6',
+                            confirmButtonText: 'Continuar'
+                        }).then((result) => {
+                            if(result.value){ location.reload(); }
+                        })
+
+                    })
+                    .catch(function (error) {
+                        // handle error
+                        console.log(error);
+                    });
+
+            }
+
+        })
+    }
 </script>
 @endsection
